split out systemd code into a new module
